+++ noodles-tabix/src/async/reader.rs
@@ -236,7 +236,7 @@ where
 {
     use crate::index::DEPTH;
 
-    let metadata_id = Bin::metadata_id(DEPTH);
+    const METADATA_ID: usize = Bin::metadata_id(DEPTH);
 
     let n_bin = reader.read_i32_le().await.and_then(|n| {
         usize::try_from(n).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
@@ -250,7 +250,7 @@ where
             usize::try_from(n).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
         })?;
 
-        if id == metadata_id {
+        if id == METADATA_ID {
             metadata = read_metadata(reader).await.map(Some)?;
         } else {
             let chunks = read_chunks(reader).await?;