+++ noodles-tabix/src/async/writer.rs
@@ -310,12 +310,11 @@ where
 {
     use crate::index::DEPTH;
 
+    const METADATA_ID: usize = Bin::metadata_id(DEPTH);
     const METADATA_CHUNK_COUNT: usize = 2;
 
-    let metadata_id = Bin::metadata_id(DEPTH);
-
     let id =
-        u32::try_from(metadata_id).map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e))?;
+        u32::try_from(METADATA_ID).map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e))?;
     writer.write_u32_le(id).await?;
 
     let n_chunk = u32::try_from(METADATA_CHUNK_COUNT)