+++ noodles-tabix/src/reader.rs
@@ -242,7 +242,7 @@ where
 {
     use super::index::DEPTH;
 
-    let metadata_id = Bin::metadata_id(DEPTH);
+    const METADATA_ID: usize = Bin::metadata_id(DEPTH);
 
     let n_bin = reader.read_i32::<LittleEndian>().and_then(|n| {
         usize::try_from(n).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
@@ -256,7 +256,7 @@ where
             usize::try_from(n).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
         })?;
 
-        if id == metadata_id {
+        if id == METADATA_ID {
             metadata = read_metadata(reader).map(Some)?;
         } else {
             let chunks = read_chunks(reader)?;